Add new DynamoDB operations
DynamoDB has added a number of new operations to their API, for exporting
tables, processing PartiQL requests, and Kinesis integration. Add these
methods as unimplemented to the DAX API so the client will build with the
latest AWS SDK (1.35.35).

Also change the behavior when unimplemented methods are called to return an
error instead of panicking, as requested in
https://github.com/aws/aws-dax-go/issues/27.